Make Dockerfile CMD robust to Ctrl+Z:
CTRL+Z will sometimes cause ade to crash.
The error occurs when a process that has subprocesses is sent to the background
with CTRL-Z. On a technical level this leads to a SIGCHLD being send to CMD, in
the case of the ade image an instance of `dash`. Per POSIX XCU 2.11 this should
not interfere with the `wait` command as no explicit userspace trap is set to
this signal, however dash's internal handling interrupts wait anyways.

To mitigate this behavior we now keep calling wait until the subprocess exits
no matter what.